cleanup, update, organised code
- updated cli to v2
- moved all cli init code out of root to a directory
- removed the need for some globals
