+++ commands/package_install.go
@@ -1,9 +1,13 @@
-package main
+package commands
 
 import (
+	"context"
+	"fmt"
+	"os"
+	"strings"
+
 	"github.com/pkg/errors"
-	"gopkg.in/segmentio/analytics-go.v3"
-	"gopkg.in/urfave/cli.v1"
+	"github.com/urfave/cli/v2"
 
 	"github.com/Southclaws/sampctl/download"
 	"github.com/Southclaws/sampctl/print"
@@ -12,20 +16,20 @@ import (
 	"github.com/Southclaws/sampctl/versioning"
 )
 
-var packageUninstallFlags = []cli.Flag{
-	cli.StringFlag{
+var PackageInstallFlags = []cli.Flag{
+	&cli.StringFlag{
 		Name:  "dir",
 		Value: ".",
 		Usage: "working directory for the project - by default, uses the current directory",
 	},
-	cli.BoolFlag{
+	&cli.BoolFlag{
 		Name:  "dev",
-		Usage: "for specifying development dependencies",
+		Usage: "for specifying dependencies only necessary for development or testing of the package",
 	},
 }
 
 //nolint:dupl
-func packageUninstall(c *cli.Context) error {
+func PackageInstall(c *cli.Context) error {
 	if c.Bool("verbose") {
 		print.SetVerbose()
 	}
@@ -33,18 +37,8 @@ func packageUninstall(c *cli.Context) error {
 	dir := util.FullPath(c.String("dir"))
 	development := c.Bool("dev")
 
-	if config.Metrics {
-		//nolint:errcheck
-		segment.Enqueue(analytics.Track{
-			Event:  "package uninstall",
-			UserId: config.UserID,
-			Properties: analytics.NewProperties().
-				Set("development", development),
-		})
-	}
-
-	if len(c.Args()) == 0 {
-		cli.ShowCommandHelpAndExit(c, "uninstall", 0)
+	if c.Args().Len() == 0 {
+		cli.ShowCommandHelpAndExit(c, "install", 0)
 		return nil
 	}
 
@@ -55,7 +49,7 @@ func packageUninstall(c *cli.Context) error {
 	}
 
 	deps := []versioning.DependencyString{}
-	for _, dep := range c.Args() {
+	for _, dep := range c.Args().Slice() {
 		deps = append(deps, versioning.DependencyString(dep))
 	}
 
@@ -64,12 +58,33 @@ func packageUninstall(c *cli.Context) error {
 		return errors.Wrap(err, "failed to interpret directory as Pawn package")
 	}
 
-	err = pcx.Uninstall(deps, development)
+	err = pcx.Install(context.Background(), deps, development)
 	if err != nil {
 		return err
 	}
 
-	print.Info("successfully removed dependency")
+	print.Info("successfully added new dependency")
 
 	return nil
+}
+
+func PackageInstallBash(c *cli.Context) {
+	cacheDir, err := download.GetCacheDir()
+	if err != nil {
+		fmt.Fprintln(os.Stderr, "Failed to retrieve cache directory path (attempted <user folder>/.samp) ", err)
+		return
+	}
+
+	packages, err := download.GetPackageList(cacheDir)
+	if err != nil {
+		fmt.Fprintln(os.Stderr, "Failed to get package list:", err)
+		return
+	}
+
+	query := c.Args().First()
+	for _, pkg := range packages {
+		if strings.HasPrefix(pkg.String(), query) {
+			fmt.Println(pkg)
+		}
+	}
 }
