+++ commands/package_run.go
@@ -1,12 +1,11 @@
-package main
+package commands
 
 import (
 	"context"
 	"os"
 
 	"github.com/pkg/errors"
-	"gopkg.in/segmentio/analytics-go.v3"
-	"gopkg.in/urfave/cli.v1"
+	"github.com/urfave/cli/v2"
 
 	"github.com/Southclaws/sampctl/download"
 	"github.com/Southclaws/sampctl/print"
@@ -14,49 +13,49 @@ import (
 	"github.com/Southclaws/sampctl/util"
 )
 
-var packageRunFlags = []cli.Flag{
-	cli.StringFlag{
+var PackageRunFlags = []cli.Flag{
+	&cli.StringFlag{
 		Name:  "dir",
 		Value: ".",
 		Usage: "working directory for the server - by default, uses the current directory",
 	},
-	cli.BoolFlag{
+	&cli.BoolFlag{
 		Name:  "container",
 		Usage: "starts the server as a Linux container instead of running it in the current directory",
 	},
-	cli.StringFlag{
+	&cli.StringFlag{
 		Name:  "build",
 		Value: "",
 		Usage: "build configuration to use if `--forceBuild` is set",
 	},
-	cli.BoolFlag{
+	&cli.BoolFlag{
 		Name:  "forceBuild",
 		Usage: "forces a build to run before executing the server",
 	},
-	cli.BoolFlag{
+	&cli.BoolFlag{
 		Name:  "forceEnsure",
 		Usage: "forces dependency ensure before build if `--forceBuild` is set",
 	},
-	cli.BoolFlag{
+	&cli.BoolFlag{
 		Name:  "noCache",
 		Usage: "forces download of plugins if `--forceEnsure` is set",
 	},
-	cli.BoolFlag{
+	&cli.BoolFlag{
 		Name:  "watch",
 		Usage: "keeps sampctl running and triggers builds whenever source files change",
 	},
-	cli.StringFlag{
+	&cli.StringFlag{
 		Name:  "buildFile",
 		Value: "",
 		Usage: "declares a file to store the incrementing build number for easy versioning",
 	},
-	cli.BoolFlag{
+	&cli.BoolFlag{
 		Name:  "relativePaths",
 		Usage: "force compiler output to use relative paths instead of absolute",
 	},
 }
 
-func packageRun(c *cli.Context) error {
+func PackageRun(c *cli.Context) error {
 	if c.Bool("verbose") {
 		print.SetVerbose()
 	}
@@ -72,23 +71,6 @@ func packageRun(c *cli.Context) error {
 	relativePaths := c.Bool("relativePaths")
 
 	runtimeName := c.Args().Get(0)
-
-	if config.Metrics {
-		//nolint:errcheck
-		segment.Enqueue(analytics.Track{
-			Event:  "package run",
-			UserId: config.UserID,
-			Properties: analytics.NewProperties().
-				Set("container", container).
-				Set("build", build).
-				Set("forceBuild", forceBuild).
-				Set("forceEnsure", forceEnsure).
-				Set("noCache", noCache).
-				Set("watch", watch).
-				Set("buildFile", buildFile != "").
-				Set("relativePaths", relativePaths),
-		})
-	}
 
 	cacheDir, err := download.GetCacheDir()
 	if err != nil {
