+++ commands/server_download.go
@@ -1,46 +1,35 @@
-package main
+package commands
 
 import (
 	appRuntime "runtime"
 
-	"gopkg.in/segmentio/analytics-go.v3"
-	"gopkg.in/urfave/cli.v1"
+	"github.com/urfave/cli/v2"
 
 	"github.com/Southclaws/sampctl/print"
 	"github.com/Southclaws/sampctl/runtime"
 	"github.com/Southclaws/sampctl/util"
 )
 
-var serverDownloadFlags = []cli.Flag{
-	cli.StringFlag{
+var ServerDownloadFlags = []cli.Flag{
+	&cli.StringFlag{
 		Name:  "version",
 		Value: "0.3.7",
 		Usage: "the SA:MP server version to use",
 	},
-	cli.StringFlag{
+	&cli.StringFlag{
 		Name:  "dir",
 		Value: ".",
 		Usage: "working directory for the server - by default, uses the current directory",
 	},
 }
 
-func serverDownload(c *cli.Context) error {
+func ServerDownload(c *cli.Context) error {
 	if c.Bool("verbose") {
 		print.SetVerbose()
 	}
 
 	version := c.String("version")
 	dir := util.FullPath(c.String("dir"))
-
-	if config.Metrics {
-		//nolint:errcheck
-		segment.Enqueue(analytics.Track{
-			Event:  "server download",
-			UserId: config.UserID,
-			Properties: analytics.NewProperties().
-				Set("version", version),
-		})
-	}
 
 	return runtime.GetServerPackage(version, dir, appRuntime.GOOS)
 }
