+++ commands/server_run.go
@@ -1,12 +1,11 @@
-package main
+package commands
 
 import (
 	"context"
 	"os"
 
 	"github.com/pkg/errors"
-	"gopkg.in/segmentio/analytics-go.v3"
-	"gopkg.in/urfave/cli.v1"
+	"github.com/urfave/cli/v2"
 
 	"github.com/Southclaws/sampctl/download"
 	"github.com/Southclaws/sampctl/print"
@@ -15,27 +14,27 @@ import (
 	"github.com/Southclaws/sampctl/util"
 )
 
-var serverRunFlags = []cli.Flag{
-	cli.StringFlag{
+var ServerRunFlags = []cli.Flag{
+	&cli.StringFlag{
 		Name:  "dir",
 		Value: ".",
 		Usage: "working directory for the server - by default, uses the current directory",
 	},
-	cli.BoolFlag{
+	&cli.BoolFlag{
 		Name:  "container",
 		Usage: "starts the server as a Linux container instead of running it in the current directory",
 	},
-	cli.BoolFlag{
+	&cli.BoolFlag{
 		Name:  "mountCache",
 		Usage: "if `--container` is set, mounts the local cache directory inside the container",
 	},
-	cli.BoolFlag{
+	&cli.BoolFlag{
 		Name:  "noCache",
 		Usage: "forces download of plugins",
 	},
 }
 
-func serverRun(c *cli.Context) error {
+func ServerRun(c *cli.Context) error {
 	if c.Bool("verbose") {
 		print.SetVerbose()
 	}
@@ -44,18 +43,6 @@ func serverRun(c *cli.Context) error {
 	container := c.Bool("container")
 	mountCache := c.Bool("mountCache")
 	noCache := c.Bool("noCache")
-
-	if config.Metrics {
-		//nolint:errcheck
-		segment.Enqueue(analytics.Track{
-			Event:  "server run",
-			UserId: config.UserID,
-			Properties: analytics.NewProperties().
-				Set("container", container).
-				Set("mountCache", mountCache).
-				Set("noCache", noCache),
-		})
-	}
 
 	cfg, err := runtime.NewConfigFromEnvironment(dir)
 	if err != nil {
